Freesurfer surface labels
Freesurfer surface labels for Thumb (D1) and Little-Finger (D5) in Brodman Area 3b of the primary somatosensory cortex S1 for the left and right hand. Data acquisition was carried out using high-resolution fMRI of S1 and data analysis with a tool entailing specificity to the cytoarchitectonics of the Brodman areas within S1. The Labels were generated using Freesurfer 6.0 and the surface group space template fsaverage.